+++ app/Http/Controllers/Api/UserController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers\Api;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use Exception;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -15,10 +17,12 @@ class UserController extends Controller
      * Display a listing of the resource.
      *
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function index()
     {
-        // stub
+        $this->authorize('viewAny', User::class);
+
         return response()->json(User::all(), Response::HTTP_OK);
     }
 
@@ -27,9 +31,12 @@ class UserController extends Controller
      *
      * @param Request $request
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
+
         // stub
         return response()->json(null, Response::HTTP_CREATED);
     }
@@ -39,10 +46,12 @@ class UserController extends Controller
      *
      * @param  User  $user
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function show(User  $user)
     {
-        // stub
+        $this->authorize('view', $user);
+
         return response()->json($user, Response::HTTP_OK);
     }
 
@@ -52,9 +61,12 @@ class UserController extends Controller
      * @param Request $request
      * @param  User  $user
      * @return JsonResponse
+     * @throws AuthorizationException
      */
     public function update(Request $request, User  $user)
     {
+        $this->authorize('update', $user);
+
         // stub
         return response()->json(null, Response::HTTP_OK);
     }
@@ -65,10 +77,12 @@ class UserController extends Controller
      * @param User $user
      * @return JsonResponse
      * @throws Exception
+     * @throws AuthorizationException
      */
     public function destroy(User  $user)
     {
-        // stub
+        $this->authorize('delete', $user);
+
         $user->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
